requirements.txt

Django==2.2.3
django-debug-toolbar==2.0
django-fontawesome==1.0
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.1

